requirements.txt
asgiref==3.4.1
astroid==2.4.2
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
dataclasses==0.6
defusedxml==0.7.1
Django==3.2.5
django-annoying==0.10.6
django-channels==0.7.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-dbbackup==3.3.0
django-debug-toolbar==3.2.2
django-filter==2.4.0
django-haystack==3.0
django-multiselectfield==0.1.12
django-quill==0.1.0
django-taggit==1.5.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
future==0.18.2
haystack==0.42
httpie==2.4.0
idna==3.2
isort==4.3.21
itypes==1.2.0
jedi==0.17.0
Jinja2==3.0.1
lazy-object-proxy==1.5.0
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
parso==0.7.0
pefile==2021.5.24
Pillow==8.3.1
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.5.2
PySDL2==0.9.7
PySocks==1.7.1
python-ptrace==0.9.8
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
